Fixed Plant Drawing
I made some changes and now plants should be drawn correctly and not have parts floating and detached from the main body. I am also taking some time to try and figure out how to add and remove visual elements in javafx so that I can remove and add plants to the display.