Sortof Working
I've taken a long while implementing a working gui with javafx. It took a while to learn what I had to do and how what I was doing was wrong and trying to figure out the correct way to implement my application. I finally figured it out and now I am ready to do some tests to test the functionality of the simulation. Currently it is stepping through the simulation on each mouse click so I can more easily catch any bugs.