Working Simulation
I realized the way I was calculating the light positions was not consistant with the way I was positioning the plants since I wasn't using the correct units for the light positions. Now I have fixed that and made the simulation run automatically with the ability to be stopped and started with a single click. Initial tests and trials have looked good , although I may make some tweaks to how the algorithm functions in the future.