Small Bug Fix
I found and the source of and removed a bug that caused the least fit plant to remain in the pool of plants to reproduce. This was the cause of the wierd behaviour that often caused a plant with 0 branches or leaves to live and pass on its bad genes.