Better Light Filtering
I changed the algorithm that handled how plants capture light to be more fair. In the previous version light was captured instantly when it detected it was inside a plant's leaf. This would mean that the light's strength was immediately halved and give an unfair advantage to one plant that had leaves overlapping with another plant. Now the light takes into account every leaf it touches at single point in time and divides up the energy given and lost evenly. Hopefully this should favor plants that spread their leaves out more.

In addition to this change I gave each leaf of a plant a cost to the plant's overall fitness. This means leaves can cost more energy than they collect and mutating it away would be beneficial.

These changes did end up negatively impacting the performance of the simulation if you are using no delays. With a greater number of leaves there is a tiny but noticeable slowdown in the time between generations.