Adding Species
I implemented a way for plants to compare if they are the same species as another plant (low genetic variation) and I created a new way for plants to reproduce based on species. Plants will only reproduce with species of their own kind. If there are no other species of that plant it will clone itself. All this was in an attempt to create more genetic diversity within the sample of plants. However, initial testing seems to indicate that most plants will still quickly converge to one species.