Weighted Plant Removal
I updated the way that plants are selected to be removed to make way for the next generation. Before I just removed the half of the population with the least fitness. Now I have implemented a weighted random selection to determine which plants must be removed. Plants with a lower fitness have a higher weight to be removed.